# api-kanban/src/main/java/fr/ryan/api_kanban/controller/UserController.java
package fr.ryan.api_kanban.controller;

import java.util.UUID;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import fr.ryan.api_kanban.dto.UpdateUserRequest;
import fr.ryan.api_kanban.dto.UserResponse;
import fr.ryan.api_kanban.service.UserService;
import jakarta.validation.Valid;

@RestController
@RequestMapping("/api/users")
public class UserController {

	private final UserService userService;

	public UserController(UserService userService) {
		this.userService = userService;
	}

	@GetMapping("/me")
	public ResponseEntity<UserResponse> me() {
		return ResponseEntity.ok(userService.getCurrentUser());
	}

	@PatchMapping("/{id}")
	public ResponseEntity<UserResponse> update(
			@PathVariable UUID id,
			@Valid @RequestBody UpdateUserRequest request) {
		return ResponseEntity.ok(userService.updateUser(id, request));
	}
}