# api-kanban/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Kanban Board API
  description: >
    Contrat de référence pour l'exercice de veille technologique backend (NestJS / Symfony / Spring Boot).
    Ce contrat est identique quel que soit le framework choisi : c'est ce qui permet de comparer les
    implémentations à la fin de l'exercice. Importez ce fichier dans Swagger Editor, Postman ou Insomnia
    pour explorer et tester les routes directement.

    Les opérations marquées [Non finalisé] (tag « Non finalisé ») ne sont pas encore implémentées
    dans le serveur : elles figurent dans la documentation pour exposer le contrat cible, pas comme
    API livrée. Un appel « Try it out » peut donc échouer (ex. 404) tant que la route n'est pas finalisée.
  version: "1.0.0"
  contact:
    name: La Plateforme — Master 1 Développement
servers:
  - url: /api
    description: local Swagger
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Récupérer le profil de l'utilisateur connecté
      responses:
        "200":
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                email: "alice@example.com"
                name: "Alice"
                role: user
                createdAt: "2026-01-15T10:00:00Z"
        "401":
          description: Authentification manquante ou token invalide/expiré
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Unauthorized
  /users/{id}:
    patch:
      tags: [Users]
      summary: Modifier les informations d'un utilisateur (y compris ses droits)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Payload invalide (ex. rôle non reconnu)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >
            Accès refusé — modification d'un autre profil sans être admin,
            ou tentative de changer un rôle sans être admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentification manquante ou invalide
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    UpdateUserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
...
    User:
      type: object
      description: >
        Représentation publique d'un utilisateur. Le champ `password` n'existe volontairement pas
        dans ce schéma : il ne doit jamais être sérialisé dans une réponse API, quelle que soit la route.
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []